// src/modules/dsfinvk/dsfinvk-schema.ts
export interface DsfinvkColumn {
  name: string;
  type: 'string' | 'numeric';
  maxLength: number | null;
}

export interface DsfinvkTable {
  file: string;
  tableName: string;
  columns: DsfinvkColumn[];
}